+++ src/angular-cesium-entities-editor/angular-cesium-entities-editor.module.ts
@@ -1,8 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AngularCesiumModule } from '../angular-cesium/angular-cesium.module';
-import { PolygonsEditorService } from './services/entity-editors/polygons-editor/polygons-editor.service';
 import { PolygonsEditorComponent } from './components/polygons-editor/polygons-editor.component';
+import { CirclesEditorComponent } from './components/circles-editor/circles-editor.component';
+import { CirclesEditorService } from './services/entity-editors/circles-editor/circles-editor.service';
+import { PolygonsEditorService } from './services/entity-editors/polygons-editor/polygons-editor.service';
 import { PolylinesEditorService } from './services/entity-editors/polyline-editor/polylines-editor.service';
 import { PolylinesEditorComponent } from './components/polylines-editor/polylines-editor.component';
 
@@ -13,14 +15,17 @@ import { PolylinesEditorComponent } from './components/polylines-editor/polyline
   ],
   declarations: [
     PolygonsEditorComponent,
+    CirclesEditorComponent,
 		PolylinesEditorComponent,
   ],
   exports: [
     PolygonsEditorComponent,
+    CirclesEditorComponent,
     PolylinesEditorComponent,
   ],
   providers: [
     PolygonsEditorService,
+    CirclesEditorService,
     PolylinesEditorService,
   ]
 })