+++ src/angular-cesium-entities-editor/components.ts
@@ -1,2 +1,3 @@
 export * from './components/polygons-editor/polygons-editor.component';
+export * from './components/circles-editor/circles-editor.component';
 export * from './components/polylines-editor/polylines-editor.component';