+++ src/angular-cesium-entities-editor/models/basic-edit-update.ts
@@ -1,0 +1,8 @@
+import { EditModes } from './edit-mode.enum';
+import { EditActions } from './edit-actions.enum';
+
+export interface BasicEditUpdate {
+  id: string;
+  editMode: EditModes;
+  editAction: EditActions;
+}