+++ src/angular-cesium-entities-editor/models/polygon-edit-options.ts
@@ -1,4 +1,5 @@
 import { PolylineEditOptions } from './polyline-edit-options';
+import { CesiumEvent } from '../../angular-cesium/services/map-events-mananger/consts/cesium-event.enum';
 
 export interface PolygonProps {
 	material?: any;
@@ -6,4 +7,6 @@ export interface PolygonProps {
 
 export interface PolygonEditOptions extends PolylineEditOptions {
 	defaultPolygonOptions?: PolygonProps;
+	dragShapeEvent?: CesiumEvent; // TODO
+	
 }